Done
Changed Schema to not require any type of address
Created the server side for the Guests with working register for the guest and CRD for the admin.
Chefs and Employees do have access to the guests information

Addon: In the users server allowed chefs and employees to access the information of a user.